+++ contracts/rewards_emission/src/contract.rs
@@ -1,8 +1,7 @@
 use cosmwasm_std::{
-    debug_print,
-    to_binary,
     Api,
     Binary,
+    debug_print,
     Env,
     Extern,
     HandleResponse,
@@ -11,15 +10,16 @@ use cosmwasm_std::{
     StdError,
     StdResult,
     Storage,
+    to_binary,
     Uint128,
 };
 
-use shade_protocol::contract_interfaces::treasury::{
-    adapter,
+use shade_protocol::contract_interfaces::dao::{
     rewards_emission::{Config, HandleMsg, InitMsg, QueryMsg},
 };
 
 use secret_toolkit::snip20::{register_receive_msg, set_viewing_key_msg};
+use shade_protocol::contract_interfaces::dao::adapter;
 
 use crate::{
     handle,
