+++ contracts/scrt_staking/src/handle.rs
@@ -1,12 +1,11 @@
 use cosmwasm_std::{
-    debug_print,
-    to_binary,
     Api,
     BalanceResponse,
     BankQuery,
     Binary,
     Coin,
     CosmosMsg,
+    debug_print,
     Env,
     Extern,
     HandleResponse,
@@ -16,6 +15,7 @@ use cosmwasm_std::{
     StdError,
     StdResult,
     Storage,
+    to_binary,
     Uint128,
     Validator,
 };
@@ -23,16 +23,15 @@ use cosmwasm_std::{
 use secret_toolkit::snip20::{deposit_msg, redeem_msg};
 
 use shade_protocol::{
-    contract_interfaces::{
-        staking::scrt_staking::{Config, HandleAnswer, ValidatorBounds},
-        treasury::{adapter, treasury::Flag},
-    },
+    contract_interfaces::dao::treasury::Flag,
     utils::{
-        asset::{scrt_balance, Contract},
+        asset::{Contract, scrt_balance},
         generic_response::ResponseStatus,
         wrap::{unwrap, wrap_and_send},
     },
 };
+use shade_protocol::contract_interfaces::dao::adapter;
+use shade_protocol::contract_interfaces::dao::scrt_staking::{Config, HandleAnswer, ValidatorBounds};
 
 use crate::{
     query,
