+++ contracts/scrt_staking/src/query.rs
@@ -15,10 +15,9 @@ use cosmwasm_std::{
     Uint128,
 };
 
-use shade_protocol::{
-    contract_interfaces::{staking::scrt_staking::QueryAnswer, treasury::adapter},
-    utils::asset::scrt_balance,
-};
+use shade_protocol::utils::asset::scrt_balance;
+use shade_protocol::contract_interfaces::dao::adapter;
+use shade_protocol::contract_interfaces::dao::scrt_staking::QueryAnswer;
 
 use crate::state::{config_r, self_address_r, unbonding_r};
