+++ contracts/treasury_manager/src/state.rs
@@ -9,7 +9,7 @@ use cosmwasm_storage::{
     ReadonlySingleton,
     Singleton,
 };
-use shade_protocol::contract_interfaces::{snip20::Snip20Asset, treasury::treasury_manager};
+use shade_protocol::contract_interfaces::{snip20::Snip20Asset, dao::treasury_manager};
 
 pub static CONFIG_KEY: &[u8] = b"config";
 pub static ASSETS: &[u8] = b"assets";
