+++ packages/shade_protocol/src/contract_interfaces/dao/DAO_ADAPTER.md
