+++ packages/shade_protocol/src/contract_interfaces/dao/adapter.rs
