+++ packages/shade_protocol/src/contract_interfaces/dao/treasury.rs
@@ -1,5 +1,4 @@
 use crate::{
-    contract_interfaces::treasury::adapter,
     utils::{asset::Contract, cycle::Cycle, generic_response::ResponseStatus},
 };
 
@@ -7,6 +6,7 @@ use cosmwasm_std::{Binary, HumanAddr, StdResult, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 use serde::{Deserialize, Serialize};
+use crate::contract_interfaces::dao::adapter;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
