+++ packages/shade_protocol/src/contract_interfaces/dao/treasury_manager.rs
@@ -1,11 +1,11 @@
 use crate::{
-    contract_interfaces::treasury::adapter,
     utils::{asset::Contract, generic_response::ResponseStatus},
 };
 use cosmwasm_std::{Binary, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 use serde::{Deserialize, Serialize};
+use crate::contract_interfaces::dao::adapter;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
