+++ packages/shade_protocol/src/contract_interfaces/staking/mod.rs
@@ -1,4 +1,2 @@
-#[cfg(feature = "scrt_staking")]
-pub mod scrt_staking;
 #[cfg(feature = "snip20_staking")]
 pub mod snip20_staking;
